requirements.txt
aiogram==3.4.1
aiofiles==23.2.1
apscheduler==3.10.4
python-dotenv==1.0.0
aiosqlite==0.19.0
yookassa==2.3.4
aiohttp==3.9.3
cryptography==42.0.5
